implement the TEXT tag
this is a tag containing an ordinary string but the tag itself has a
`TI` (text identifier) attribute. The only user is the `DTC` tag and
it also seems to be very unusual to specify a text identifier there.

This is the last is missing tag type that can be specified inside
DIAG-LAYER-CONTAINER, COMPARAM-SUBSET or COMPARAM-SPEC, i.e., after
this commit the diagnostics related functionality of the ODX standard
should be fully covered by odxtools... (modulo bugs, obviously.)

Signed-off-by: Andreas Lauser <andreas.lauser@mercedes-benz.com>
Signed-off-by: Christian Hackenbeck <christian.hackenbeck@mercedes-benz.com>