Put classes representing XML tag types into separate files
The purpose of this is to make the codebase consistently sticking to
the "one python file per tag type in the XSD" mantra. Unfortunately,
this is also a pretty long and boring patch, but at least it changes
absolutely nothing from the functional perspective so it should be
relatively easy to review.

hopefully this is going be the last janitorial commit for a while...

Signed-off-by: Andreas Lauser <andreas.lauser@mercedes-benz.com>
Signed-off-by: Christian Hackenbeck <christian.hackenbeck@mercedes-benz.com>