unify the bitmask output code
the standard does not seem to allow an odd number of nibbles (i.e.,
the examples featuring an odd number of nibbles do not pass XML
validation using the XSD), so we use the same code for printing bit
masks for `POS-RESPONSE-SUPPRESSABLE` as for diag services.

this patch was inspired by one of [at]kayoub5's comments; thanks!

Signed-off-by: Andreas Lauser <andreas.lauser@mercedes-benz.com>