generalize ODXLINK reference writing
writing ODXLINK references should now always work regardless of in
which document fragments the reference and the target objects are
located.

this works by providing three new "magical" global functions to the
jinja macros for output:

- `set_category_docfrag()`: Specify the document fragment for the ODX
  category which is applicable for all subsequent macro invocations.
- `set_layer_docfrag()`: Specify the document fragment for the
  diagnostic layer which is applicable for all subsequent macro
  invocations (for DIAG-LAYER-CONTAINER).
- `make_ref_attribs(ref: OdxLinkRef)`: produces all necessary XML
  attributes required for an ODX link reference XML tag. (i.e.,
  ID-REF, DOCREF and DOCTYPE)

internally, these functions require a dictionary parameter for storing
"magic" global variables. On the jinja side, this is not necessary
because lambda functions are used as wrappers.

Signed-off-by: Andreas Lauser <andreas.lauser@mercedes-benz.com>
Approved-by: Alexander Walz <alexander.walz@mercedes-benz.com>