add a `strip_indent()` utility function
this function removes the leading and trailing space characters of
each lines of a string. This is useful for not letting the spaces
added by the output writing templates modify the resulting database
object. Note that this function only removes space and carriage-return
characters, so other white space like tabulators are unaffected.

A better solution would be to instruct Jinja that a given part of the
output should be unaffected by `indent` filters, but unfortunately
this is impossible as far as I can see...

Signed-off-by: Andreas Lauser <andreas.lauser@mercedes-benz.com>