´.odx-v': implement support for internalizing `VEHICLE-INFORMATION-SPEC`
`VEHICLE-INFORMATION-SPEC` is used to describe how a given vehicle can
be identified in a semi-automatic way. Also, it describes the
vehicle's network topology, i.e., how communication with the vehicle's
ECUs can be achieved for networking architectures involving
gateways. This is described in section 7.3.10 of the ASAM version of
the standard.

Signed-off-by: Andreas Lauser <andreas.lauser@mercedes-benz.com>
Approved-by: Michael Hahn <michael.a.hahn@mercedes-benz.com>